requirements.txt
haystack-ai~=2.16.1
chroma-haystack~=3.3.0
numpy<2
ollama-haystack~=4.1.0
google-genai-haystack~=2.1.1
mcp-haystack~=0.4.0
chardet~=5.2.0
sentence-transformers~=5.1.0
prompt_toolkit~=3.0.51
mcp[cli]~=1.12.4
httpx~=0.28.1
uv~=0.8.6
tldextract~=5.3.0
validators~=0.35.0
more-itertools~=10.7.0
requests~=2.32.4
beautifulsoup4~=4.13.4
aiofiles~=24.1.0
bitarray~=3.5.1
persist-queue~=1.0.0
mitmproxy~=12.1.1
lxml~=6.0.0
tinycss2~=1.4.0
json5~=0.12.0
html5lib~=1.1
optimum~=1.27.0
ddgs~=9.5.1
